import/pyproject.toml
[project]
name = "numad-import"
version = "0.1.0"
description = "Numad data import"
readme = "README.md"
authors = [{ name = "Miel Peeters", email = "peetersmiel@gmail.com" }]
requires-python = "~=3.13.0"
dependencies = [
    "geoalchemy2>=0.18.1",
    "psycopg2-binary>=2.9.11",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlmodel>=0.0.27",
]

[dependency-groups]
dev = ["eralchemy[graphviz]>=1.6.0"]

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
import = "numad_import.cli:main"
erd = "numad_import.cli:erd"
